feat(jwt.util): Create jwToken sign and decode util
